(fix): resolve incorrect and inconsistent module naming
- see https://github.com/agilgur5/django-serializable-model/issues/2
  - it had the same issue and its changes were cherry-picked here
- in Python-land, packages are _not_ imported by their package name,
  but their _module_ name
  - see also https://stackoverflow.com/a/54599368/3431180
  - and the module name of this package was `decorators`, _not_
    `django-api-decorators`
- also in Python-land, one cannot use the standard `import` syntax
  for hyphenated modules, it will actually cause a SyntaxError
  - but for some reason it's not illegal to do so, and is allowed both
    in `setuptools` naming and PyPI naming (it is discouraged by PEP8)

- rename decorators.py to django_api_decorators.py
  - leave decorators.py and just make it a wrapper / alias of
    django_serialiazable_model so as not to cause a breaking change

(pkg): publish `django_api_decorators` module as well
- don't remove `decorators` though so as to not cause a breaking
  change; should remove it in v1.0.0

(docs): use `django_api_decorators` in Usage docs
